+++ src/components/DataView/EmptyState.js
@@ -1,8 +1,112 @@
-import React from 'react'
+import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
+import { useTheme } from '../../theme'
 import { GU, textStyle } from '../../style'
+import { usePublicUrl } from '../../providers/PublicUrl'
+import LoadingRing from '../LoadingRing/LoadingRing'
+import Link from '../Link/Link'
+import illustrationRedImage from './assets/empty-state-illustration-red.png'
+import illustrationBlueImage from './assets/empty-state-illustration-blue.png'
 
-function EmptyState({ status, emptyState }) {
+function prepareDefaultStates(
+  publicUrl,
+  statusEmpty,
+  statusLoading,
+  statusEmptyFilters,
+  statusEmptySearch
+) {
+  return {
+    default: {
+      illustration: publicUrl + illustrationBlueImage,
+      description: statusEmpty || 'No data available.',
+    },
+    loading: {
+      illustration: publicUrl + illustrationBlueImage,
+      description: statusLoading || 'Loading data…',
+    },
+    'empty-filters': {
+      illustration: publicUrl + illustrationRedImage,
+      description:
+        statusEmptyFilters ||
+        'We can’t find any item matching your filter selection. ',
+    },
+    'empty-search': {
+      illustration: publicUrl + illustrationRedImage,
+      description:
+        statusEmptySearch ||
+        'We can’t find any item matching your search query. ',
+    },
+  }
+}
+
+function useEmptyStateValue(
+  status,
+  defaultEmptyStates,
+  emptyStateConfigurator
+) {
+  if (!emptyStateConfigurator) {
+    return defaultEmptyStates[status]
+  }
+
+  const functionMode = typeof emptyStateConfigurator === 'function'
+  const customEmptyState = functionMode
+    ? emptyStateConfigurator(status)
+    : emptyStateConfigurator[status]
+
+  // description shortcut for object mode with a node
+  if (!functionMode && React.isValidElement(customEmptyState)) {
+    return { ...defaultEmptyStates[status], description: customEmptyState }
+  }
+
+  // description shortcut for object mode with a string
+  if (!functionMode && typeof customEmptyState === 'string') {
+    return { ...defaultEmptyStates[status], description: customEmptyState }
+  }
+
+  // override all (illustration and text) with function mode
+  if (React.isValidElement(customEmptyState)) {
+    return customEmptyState
+  }
+
+  return {
+    ...defaultEmptyStates[status],
+    ...customEmptyState,
+  }
+}
+
+function EmptyState({
+  status,
+  emptyStateConfigurator,
+  statusEmpty,
+  statusLoading,
+  statusEmptyFilters,
+  statusEmptySearch,
+  onStatusEmptyClear,
+}) {
+  const publicUrl = usePublicUrl()
+  const theme = useTheme()
+  const defaultEmptyStates = useMemo(() => {
+    return prepareDefaultStates(
+      publicUrl,
+      statusEmpty,
+      statusLoading,
+      statusEmptyFilters,
+      statusEmptySearch
+    )
+  }, [
+    publicUrl,
+    statusEmpty,
+    statusEmptyFilters,
+    statusEmptySearch,
+    statusLoading,
+  ])
+
+  const emptyState = useEmptyStateValue(
+    status,
+    defaultEmptyStates,
+    emptyStateConfigurator
+  )
+
   if (React.isValidElement(emptyState)) {
     return emptyState
   }
@@ -41,6 +145,35 @@ function EmptyState({ status, emptyState }) {
 
         {(() => {
           // Empty state: content part
+          if (status === 'default') {
+            return (
+              <p
+                css={`
+                  ${textStyle('title2')};
+                `}
+              >
+                {emptyState.description}
+              </p>
+            )
+          }
+          if (status === 'loading') {
+            return (
+              <p
+                css={`
+                  ${textStyle('title2')};
+                  display: flex;
+                  align-items: center;
+                `}
+              >
+                <LoadingRing
+                  css={`
+                    margin-right: ${2 * GU}px;
+                  `}
+                />
+                {emptyState.description}
+              </p>
+            )
+          }
           if (status === 'empty-filters' || status === 'empty-search') {
             return (
               <>
@@ -52,7 +185,15 @@ function EmptyState({ status, emptyState }) {
                 >
                   No results found.
                 </p>
-                {emptyState.description}
+
+                <p
+                  css={`
+                    color: ${theme.surfaceContentSecondary};
+                  `}
+                >
+                  {emptyState.description}
+                  <Link onClick={onStatusEmptyClear}>Clear filters</Link>
+                </p>
               </>
             )
           }
@@ -70,7 +211,12 @@ EmptyState.propTypes = {
     'empty-filters',
     'empty-search',
   ]),
-  emptyState: PropTypes.object,
+  emptyStateConfigurator: PropTypes.object,
+  statusEmpty: PropTypes.node,
+  statusLoading: PropTypes.node,
+  statusEmptyFilters: PropTypes.node,
+  statusEmptySearch: PropTypes.node,
+  onStatusEmptyClear: PropTypes.func,
 }
 
 export default EmptyState
